major refactor for v5
