+++ composer.json
@@ -21,6 +21,9 @@
     "composer/installers": "~1.0",
     "erusev/parsedown": "^1.7"
   },
+  "config": {
+    "vendor-dir": "src/lib"
+  },
   "autoload": {
     "classmap": [
       "/"
