+++ src/inc/class-panel.php
