+++ src/inc/parts/class-color.php
