+++ src/inc/parts/class-email.php
