+++ src/inc/parts/class-markdown.php
