+++ src/inc/parts/class-password.php
