+++ src/inc/parts/class-text.php
