+++ src/inc/parts/class-toggle-switch.php
